pyproject.toml
[project]
name = "pdf-to-markdown"
version = "0.1.0"
description = ""
authors = [
    {name = "JokerQianwei",email = "1174665346@qq.com"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "pymupdf (>=1.23.8,<2.0.0)",
    "google-generativeai (>=0.3.2,<0.4.0)",
    "pillow (>=10.1.0,<11.0.0)",
    "click (>=8.1.7,<9.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.14.1"
pytest-cov = "^6.1.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pdf2md = "pdf_to_markdown.cli:main"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
